refactor: Se separa el modal de buscar el cliente en un componente separado
